% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_clusters.R
\docType{data}
\name{classes_clusters}
\alias{classes_clusters}
\title{Species Classification}
\format{
A data frame with 213 rows and 120 columns:
\describe{
\item{species}{species name}
\item{Forest_Sp}{occurrence of forest species, binary (0/1)}
\item{Shrub_Sp}{occurrence of shrubland species, binary (0/1)}
\item{Grassland_Sp}{occurrence of grassland species, binary (0/1)}
\item{generalists_Sp}{occurrence of generalist species, binary (0/1)}
\item{other_specialists_Sp}{occurrence of other specialist species, binary (0/1)}
}
}
\source{
Original data from
}
\usage{
classes_clusters
}
\description{
An data set containing binary species occurrence data and coordinate information
}
\keyword{datasets}
