% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rasterfile.R
\name{load_rasterfile}
\alias{load_rasterfile}
\title{Load Example Data SpatRaster}
\usage{
load_rasterfile(force = FALSE)
}
\arguments{
\item{force}{Logical value. If \code{force = TRUE}, force re-download even if the file already exists in the cache. \code{FALSE} by default.}
}
\value{
A SpatRaster object (from the \code{terra} package) containing the 1995 land use/land cover classification with 20m resolution and EPSG:3763.
}
\description{
Downloads land use raster "coa95_raster_20.tif", which contains data from 1995, from the GitHub release of the csarGeo package. It caches the raster locally to prevent re-downloading.
}
\examples{
\dontrun{
# Load the raster (downloads if not cached)
land_use <- load_rasterfile()

# Force re-download
land_use <- load_rasterfile(force = TRUE)
}
}
