% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data.R
\docType{data}
\name{species_data}
\alias{species_data}
\title{Species Data}
\format{
A data frame with 213 rows and 120 columns:
\describe{
\item{locationID}{Site identifier}
\item{long}{Longitude coordinate}
\item{lat}{Latitude coordinate}
\item{...}{Additional 117 columns containing species presence/absence data (0/1)}
}
}
\source{
Original data from
}
\usage{
species_data
}
\description{
An dataset containing binary species occurrence data and coordinate information
}
\keyword{datasets}
