% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals_sar.R
\name{visual_sar}
\alias{visual_sar}
\title{Visualize countryside_sar results}
\usage{
visual_sar(
  result,
  plot_type = "both",
  habitat_raster = NULL,
  cluster_level = "size_256",
  plot_all_levels = FALSE,
  ncol = NULL,
  point_size = 16,
  point_col = "black",
  hull_col = "red",
  area_unit = "m",
  plot_runs = FALSE,
  show_average = TRUE,
  average_color = "red",
  average_lwd = 3,
  show_formula = TRUE,
  show_slope = TRUE,
  show_intercept = FALSE,
  show_p_value = FALSE,
  show_r_squared = TRUE,
  show_adj_r_squared = FALSE,
  show_run_legend = TRUE
)
}
\arguments{
\item{result}{Output object from countryside_sar function.}

\item{plot_type}{Type of plot: "spatial" for sampling design, "sar" for species-area relationship, or "both" in a grid.}

\item{habitat_raster}{Optional parameter to add a land-use raster background.}

\item{cluster_level}{For clusters method, defines which level to plot (e.g., "size_256")}

\item{plot_all_levels}{For "clusters" method, plot all levels at once in an arranged grid. If plot_all_levels = TRUE it is not possible to define a value for cluster_level.}

\item{ncol}{description}

\item{point_size}{Optional parameter to change the point size of the plot.}

\item{point_col}{Optional parameter to change the colour of the points.}

\item{hull_col}{Optional parameter to change the hull colour of the plot.}

\item{area_unit}{description}

\item{plot_runs}{description}

\item{show_average}{description}

\item{average_color}{description}

\item{average_lwd}{description}

\item{show_formula}{description}

\item{show_slope}{description}

\item{show_intercept}{description}

\item{show_p_value}{description}

\item{show_r_squared}{description}

\item{show_adj_r_squared}{description}

\item{show_run_legend}{description}
}
\value{
No return value, function creates plots.
}
\description{
A function to visualize the results of the function countryside_sar. It can create a "spatial" plot for the sampling process of either method and a "sar" plot for the results of the SAR-analysis and linear model results.
}
\examples{
\dontrun{
plot_countryside_sar(res, plot_type = "both")
}
}
