% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals_sar.R
\name{visuals_sar}
\alias{visuals_sar}
\title{Visualize countryside_sar results}
\usage{
visuals_sar(
  result,
  plot_type = NULL,
  plot_all_runs = TRUE,
  plot_run_n = NULL,
  plot_all_levels = TRUE,
  plot_level_n = NULL
)
}
\arguments{
\item{result}{Output object from countryside_sar function.}

\item{plot_type}{Type of plot: "map" to create a map of the polygon hull + sampling locations with the chosen sampling method, "sar" for a simple log-log scaled linear regression and "csar" for a heatmap of species group affinities to habitat types. "map" and  "sar" are available for results of both "circles" and "clusters", "csar" is only available for results of "clusters".}

\item{plot_all_runs}{Plot all runs of "circles" into one grid. Defaults to TRUE.}

\item{plot_run_n}{Plot only the result of the n-th run of "circles". Defaults to NULL.}

\item{plot_all_levels}{Plot all clustering levels of method "clusters" into one grid. Defaults to TRUE.}

\item{plot_level_n}{Plot only the n-th clustering level as a solo plot. Defaults to NULL.}
}
\value{
Creates either a single plot or a grid of multiple plots.
}
\description{
A function to visualize the results of countryside_sar() from the csarGeo-Package. It can generate three plot types: "map", a plot of the polygon hull with the sampling locations and the chosen sampling method, "sar" for the base SAR analysis and "csar", only available for a result of method "clusters" to create an affinity heat map of the species groups to different habitats.
}
\examples{
\dontrun{
plot_countryside_sar(res, plot_type = "map")
}
}
